backend/requirements.txt
fastapi
uvicorn
requests
PyMuPDF
beautifulsoup4
google-generativeai 